Add container universe support
This allows the executor to specify an HTCondor universe, and a container image when the
universe is set to container/docker. These values are passed to the executor via the
executor's CLI, e.g.:
`snakemake --executor htcondor --htcondor-universe container --htcondor-container-image docker://foo:latest

I chose to add these to the CLI instead of to the job's resources dict, because I think these will likely
be applied to all jobs in a workflow.